refactor: svg 컴포넌트로 분리, 배럴패턴 구현 feature/TODAK-123/update-diary
