+++ components/svg/DirectionSVG.tsx
@@ -1,0 +1,27 @@
+"use client";
+
+import React from "react";
+
+const DirectionSVG: React.FC = () => {
+  return (
+    <svg
+      className="w-5 h-5"
+      aria-hidden="true"
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      fill="none"
+      viewBox="0 0 24 24"
+    >
+      <path
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="2"
+        d="m15 19-7-7 7-7"
+      />
+    </svg>
+  );
+};
+
+export default DirectionSVG;
