+++ components/svg/index.ts
@@ -1,0 +1,6 @@
+import HomeSVG from "./HomeSVG";
+import PencilSVG from "./PencilSVG";
+import UserSVG from "./UserSVG";
+import DirectionSVG from "./DirectionSVG";
+
+export { HomeSVG, PencilSVG, UserSVG, DirectionSVG };
